Fix possible bug regarding passing empty strings to $wrapAround
Unknowingly passing an empty string to ansiEscape()'s $wrapAround would
turn off wrap around behaviour entirely, causing the open-ended
behaviour to take effect instead, which was very unlikely to be the
desired behaviour.

This changes things to strictly check $wrapAround to be either 'null'
or 'false' (via ===), so that it's unlikely to be unintended behaviour.

And, changes the default values of $wrapAround so that $wrapAround is
still the non-default behaviour.